+++ packages/container/angel_container/lib/angel3_container.dart
@@ -6,3 +6,4 @@ export 'src/static/static.dart';
 export 'src/exception.dart';
 export 'src/reflector.dart';
 export 'src/throwing.dart';
+export 'src/container_const.dart';