+++ packages/container/angel_container/lib/src/container_const.dart
@@ -1,0 +1,8 @@
+class ContainerConst {
+  static const String defaultErrorMessage =
+      'You attempted to perform a reflective action, but you are using `ThrowingReflector`, '
+      'a class which disables reflection. Consider using the `MirrorsReflector` '
+      'class if you need reflection.';
+
+  ContainerConst._();
+}